controller/webserver/requirements.txt
alembic==1.0.11
blessings==1.7
certifi==2019.6.16
chardet==3.0.4
Click==7.0
curtsies==0.3.0
Flask==1.0.3
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
Pygments==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.5
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4